Unit tests for formatColHistory()
