+++ app/cogs/colours/cmd_colinfo_test.go
@@ -1,6 +1,7 @@
 package colours
 
 import (
+	"strings"
 	"testing"
 	"time"
 
@@ -8,10 +9,15 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+func unix(n int64) time.Time {
+	return time.Unix(n, 0)
+}
+
+func concat(s ...string) string {
+	return strings.Join(s, "")
+}
+
 func Test_partitionColours(t *testing.T) {
-	unix := func(n int64) time.Time {
-		return time.Unix(n, 0)
-	}
 	colours := func(hxs ...int) []*Colour {
 		ret := make([]*Colour, len(hxs))
 		for i, hx := range hxs {
@@ -85,4 +91,111 @@ func Test_partitionColours(t *testing.T) {
 			)
 		})
 	}
+}
+
+func Test_formatColInfo(t *testing.T) {
+	now := time.Unix(1682000000, 0)
+
+	testCases := []struct {
+		desc            string
+		rerollCDEndTime time.Time
+		lastMutateTime  time.Time
+		lastFrozenTime  time.Time
+		history         *History
+
+		expectDesc string
+	}{
+		{
+			desc:            "reroll CD elapsed",
+			rerollCDEndTime: now.Add(1 * time.Hour),
+			expectDesc: concat(
+				"**Reroll cooldown:**\n1 hour\n\n",
+				"**Last mutate:**\n_(Never)_",
+			),
+		},
+		{
+			desc:           "have mutated, 1 hour before",
+			lastMutateTime: now.Add(-1 * time.Hour),
+			expectDesc: concat(
+				"**Reroll cooldown:**\n_(No cooldown, reroll available)_\n\n",
+				"**Last mutate:**\n1 hour ago",
+			),
+		},
+		{
+			desc:           "have mutated, exactly now",
+			lastMutateTime: now,
+			expectDesc: concat(
+				"**Reroll cooldown:**\n_(No cooldown, reroll available)_\n\n",
+				"**Last mutate:**\nMoments ago",
+			),
+		},
+		{
+			desc:           "have mutated, but frozen",
+			lastMutateTime: now.Add(-1 * time.Hour),
+			lastFrozenTime: now.Add(-1 * time.Hour),
+			expectDesc: concat(
+				"**Reroll cooldown:**\n_(No cooldown, reroll available)_\n\n",
+				"**Last mutate:**\n1 hour ago\n",
+				"Frozen 1 hour ago",
+			),
+		},
+		{
+			desc:           "never mutated, but frozen",
+			lastMutateTime: Never,
+			lastFrozenTime: now.Add(-1 * time.Hour),
+			expectDesc: concat(
+				"**Reroll cooldown:**\n_(No cooldown, reroll available)_\n\n",
+				"**Last mutate:**\n_(Never)_\n",
+				"Frozen 1 hour ago",
+			),
+		},
+		{
+			desc:           "have mutated, have history",
+			lastMutateTime: now.Add(-1 * time.Hour),
+			history: &History{
+				start: unix(00),
+				end:   unix(40),
+				records: []*ColoursLogRecord{
+					{TStamp: unix(00), ColourHex: "ffffff"},
+					{TStamp: unix(20), ColourHex: "000000"},
+				},
+			},
+			expectDesc: concat(
+				"**Reroll cooldown:**\n_(No cooldown, reroll available)_\n\n",
+				"**Last mutate:**\n1 hour ago\n\n",
+				"**Image history, newest → oldest:**",
+			),
+		},
+	}
+	cog := &Cog{cfg: &Config{MutateCooldownMins: 1}}
+	for _, tc := range testCases {
+		t.Run(tc.desc, func(t *testing.T) {
+			reply, err := cog.formatColInfo(
+				now,
+				tc.rerollCDEndTime,
+				tc.lastMutateTime,
+				tc.lastFrozenTime,
+				tc.history,
+			)
+			assert.NoError(t, err)
+			assert.Equal(t, tc.expectDesc, reply.desc)
+		})
+	}
+}
+
+func Test_formatColHistory(t *testing.T) {
+	h := &History{
+		start: unix(00),
+		end:   unix(40),
+		records: []*ColoursLogRecord{
+			{TStamp: unix(00), ColourHex: "ffffff"},
+			{TStamp: unix(20), ColourHex: "000000"},
+		},
+	}
+
+	file, fileExt, fileContent, err := formatColHistory(h, 20*time.Second)
+	assert.NoError(t, err)
+	assert.NotNil(t, file)
+	assert.Equal(t, "png", fileExt)
+	assert.Equal(t, "image/png", fileContent)
 }
